refactor identitysearch
